# Conflicts:
#	public/css/common.css
#	resources/views/auth/login.blade.php

+++ routes/web.php
@@ -11,60 +11,45 @@
 |
 */
 
-
+// ①topページ
+Route::get('/', 'HomeController@index')->name('home.index');//home画面へ
 Route::get('/event', 'EventController@index')->name('event.index');//一覧表示 ここの場合->name('event.index');はなくてもOK でも今後こっちの方が便利になるかもしれないから書いている
 Route::get('/event/search', 'EventController@search')->name('event.search'); //->以降のコードはindex.blade.phpのアクションで指名してもらえるように同じ名前をかく
 // '/event/search'は別に画面遷移するというわけではなく同じページでもURLは違っていてもOK
-Route::get('/sakagura', 'SakaguraController@index')->name('sakagura.index');
-// Route::get('/event', 'EventController@index')->name('event.index');
-Route::get('/sakagura/mypage', 'SakaguraController@mypage')->name('sakagura.mypage');
-
-
-// 酒蔵新規登録・ログイン用
-Route::group(['prefix' => 'sakagura', 'middleware' => 'guest:owner'], function() {
-    // Route::get('/', function () {
-    //     return view('sakagura');
-    // });
 
 
 
+// ②酒蔵ログインしてる場合だけ行える処理
+Route::group(['prefix' => 'owner', 'middleware' => 'auth:owner'], function () {
+    Route::get('/', 'Owner\OwnerController@mypage')->name('owner.mypage');
+    Route::post('logout', 'Owner\Auth\LoginController@logout')->name('owner.logout');
+    Route::get('event/create', 'Owner\EventController@create')->name('event.create');//イベント登録
+    Route::post('event/create', 'Owner\EventController@store')->name('event.store');//イベント保存
+    Route::delete('event/{event}/delete', 'Owner\EventController@destroy')->name('event.destroy'); // 削除処理
+    Route::get('event/{event}/edit', 'Owner\EventController@edit')->name('event.edit'); // 編集画面
+    Route::put('event/{event}/update', 'Owner\EventController@update')->name('event.update'); // 更新処理
+});
+
+
+
+// ③酒蔵ログインしてない場合だけ行える処理
+Route::group(['prefix' => 'owner', 'middleware' => 'guest:owner'], function () {
     Route::get('login', 'Owner\Auth\LoginController@showLoginForm')->name('owner.login');
     Route::post('login', 'Owner\Auth\LoginController@login')->name('owner.login');
-
     Route::get('register', 'Owner\Auth\RegisterController@showRegisterForm')->name('owner.register');
-    Route::post('register', 'Owner\Auth\RegisterController@create')->name('owner.register');
-
-    // Route::get('password/rest', 'Owner\Auth\ForgotPasswordController@showLinkRequestForm')->name('owner.password.request');
+    Route::post('register', 'Owner\Auth\RegisterController@register')->name('owner.register');
 });
 
-// 酒蔵ログアウト用
-Route::group(['prefix' => 'sakagura', 'middleware' => 'auth:owner'], function(){
-
-    Route::get('mypage', 'SakaguraController@mypage')->name('sakagura.mypage');
-    Route::post('logout', 'Owner\Auth\LoginController@logout')->name('owner.logout');
 
 
-    Route::get('event/create', 'EventController@event_create')->name('event.create');//イベント登録
-    Route::post('event/create', 'EventController@event_store')->name('event.store');//イベント保存
-    Route::delete('event/{event}/delete', 'EventController@destroy')->name('event.destroy'); // 削除処理
-
-    Route::get('event/{event}/edit', 'EventController@event_edit')->name('event.edit'); // 編集画面
-    Route::put('event/{event}/update', 'EventController@event_update')->name('event.update'); // 更新処理
-
-
-    // Route::get('home', 'Owner\HomeController@index')->name('owner.home');
-
+// ④外国人ユーザーがログインしてる場合だけ行える処理
+Route::group(['middleware' => 'auth'], function () {
+    Route::get('event/{id}/apply', 'EventController@apply')->name('event.apply'); // 申込画面
+    Route::put('event/{id}/applyed', 'EventController@applyed')->name('event.applyed'); // 申込更新処理
+    Route::delete('event/{id}/applydelete', 'EventController@cancel')->name('event.applydestroy'); // 申込削除処理
 });
 
-// 酒蔵側の認証機能
-// 以下のページはログインしているときのみ表示
-// Route::get('/owner/mypage', 'Owner\Auth\OwnerMyPageController@index')->name('owner.mypage');
-// Route::group(['prefix' => 'owner','middleware' => 'auth'], function() {
-//     Route::get('owner/mypage', 'Owner\Auth\OwnerMyPageController@index')->name('owner.mypage');
-// });
-                                // 👆はOwnerMypageController.phpのnamespace App\Http\Controllers\Owner\Auth;と同じ事をかく ここでlaravelさんに辿りついてもらうために書かないといけない じゃないと見つけてもらえなくなる
 
+
+// ⑤外国人ユーザー認証関連
 Auth::routes();
-Route::get('event/{id}/apply', 'EventController@event_apply')->name('event.apply'); // 申込画面
-Route::put('event/{id}/applyed', 'EventController@event_applyed')->name('event.applyed'); // 申込更新処理
-Route::delete('event/{id}/applydelete', 'EventController@event_applydestroy')->name('event.applydestroy'); // 申込削除処理
